separate warnings module into partial classes
